Player Lobby
This commit has all the initial goodies from making cross-client communication
work. It comes in the form of a player lobby scene wherein the player's names
are displayed.

While waiting for the second player, the text "WAITING..." is displayed but
is updated with the second player's name once they join. To achieve this, I've
had to add a whole bunch of stuff:

* Dynamic text components
  A new component that can change the content of a text component via messages

* Horizontal layouts
  A new component that can lay out components horizontally

* Entities
  The concept of entities as the place where communication across multiple
  components is bridged.

* State as a tea.Model
  The game state is now pumped into each client's loop and will handle
  messages it cares about

* Containers
  A new Container component that just wraps another component and applies
  styles.

Signed-off-by: David Bond <davidsbond93@gmail.com>